Cutover all details tabs to v4 metrics and fixed composers of filters for automatic membership

+++ supabase/seeds/demo-tenants/01_electricals.sql
@@ -92,7 +92,7 @@ BEGIN
   -- 7. Tenant products
   INSERT INTO app.tenant_products (tenant_id, tenant_brand_id, master_product_id, internal_sku, mrp, base_selling_price, cost_price, is_active)
   SELECT v_tenant_id, tb.id, p.id, p.master_sku, v.mrp, v.base_rate,
-         round(v.base_rate * (1 - (0.05 + random() * 0.15)), 2), true
+         round((v.base_rate * (1 - (0.05 + random() * 0.15)))::numeric, 2), true
   FROM (VALUES
     ('ELE-PLC-WIRE15', 1610, 1420),
     ('ELE-PLC-WIRE25', 2480, 2190),
