+++ supabase/seeds/demo-tenants/02_mobiles_electronics.sql
@@ -81,7 +81,7 @@ BEGIN
 
   INSERT INTO app.tenant_products (tenant_id, tenant_brand_id, master_product_id, internal_sku, mrp, base_selling_price, cost_price, is_active)
   SELECT v_tenant_id, tb.id, p.id, p.master_sku, v.mrp, v.base_rate,
-         round(v.base_rate * (1 - (0.05 + random() * 0.15)), 2), true
+         round((v.base_rate * (1 - (0.05 + random() * 0.15)))::numeric, 2), true
   FROM (VALUES
     ('MOB-SAM-A16-128', 18999, 17200),
     ('MOB-SAM-M35-128', 19999, 18100),
